chore: upgrade svelte, sveltekit, tailwindcss, authjs, vite
