+++ package.json
@@ -34,33 +34,34 @@
 	},
 	"homepage": "https://github.com/conrmahr/discord-rss",
 	"devDependencies": {
-		"@sveltejs/adapter-node": "^3.0.1",
-		"@sveltejs/kit": "^2.5.27",
-		"@sveltejs/vite-plugin-svelte": "^4.0.0",
-		"@tailwindcss/forms": "^0.5.7",
-		"@types/node": "^20.9.1",
-		"@typescript-eslint/eslint-plugin": "^6.10.0",
-		"@typescript-eslint/parser": "^6.10.0",
-		"autoprefixer": "^10.4.17",
-		"dotenv": "^16.3.1",
-		"eslint": "^8.28.0",
-		"eslint-config-prettier": "^9.0.0",
-		"eslint-plugin-svelte": "^2.45.1",
-		"postcss": "^8.4.33",
-		"prettier": "^3.1.0",
-		"prettier-plugin-svelte": "^3.2.6",
-		"svelte": "^5.0.0",
-		"svelte-check": "^4.0.0",
-		"tailwindcss": "^3.3.5",
-		"tslib": "^2.4.1",
-		"typescript": "^5.5.0",
-		"vite": "^5.4.4"
+		"@sveltejs/adapter-node": "^5.2.13",
+		"@sveltejs/kit": "^2.27.0",
+		"@sveltejs/vite-plugin-svelte": "^6.1.0",
+		"@tailwindcss/forms": "^0.5.10",
+		"@types/node": "^24.2.0",
+		"@typescript-eslint/eslint-plugin": "^8.39.0",
+		"@typescript-eslint/parser": "^8.39.0",
+		"dotenv": "^17.2.1",
+		"eslint": "^9.32.0",
+		"eslint-config-prettier": "^10.1.8",
+		"eslint-plugin-svelte": "^3.11.0",
+		"postcss": "^8.5.6",
+		"prettier": "^3.6.2",
+		"prettier-plugin-svelte": "^3.4.0",
+		"svelte": "^5.37.3",
+		"svelte-check": "^4.3.1",
+		"tailwindcss": "^4.1.11",
+		"tslib": "^2.8.1",
+		"typescript": "^5.9.2",
+		"vite": "^7.0.6"
 	},
 	"type": "module",
 	"dependencies": {
-		"@auth/core": "^0.21.0",
-		"@auth/sveltekit": "^0.7.0",
-		"ioredis": "^5.3.2",
+		"@auth/core": "^0.40.0",
+		"@auth/sveltekit": "^1.10.0",
+		"@tailwindcss/postcss": "^4.1.11",
+		"@tailwindcss/vite": "^4.1.11",
+		"ioredis": "^5.7.0",
 		"rss-parser": "^3.13.0"
 	},
 	"engines": {
