+++ src/routes/+layout.server.ts
@@ -15,7 +15,7 @@ const getFeeds = async () => {
 
 // past current feeds to session
 export const load: LayoutServerLoad = async (event) => {
-	const session = await event.locals.getSession();
+	const session = await event.locals.auth();
 	return {
 		session: session,
 		get: await getFeeds()
