+++ src/routes/+page.svelte
@@ -258,7 +258,7 @@
 									</td>
 									<td class="whitespace-nowrap py-5 pl-4 pr-3 text-sm sm:pl-0">
 										<div class="flex items-center">
-											<div class="h-11 w-11 flex-shrink-0">
+											<div class="h-11 w-11 shrink-0">
 												<img class="h-11 w-11 rounded-full" src={sub.thumbnail} alt="" />
 											</div>
 											<div class="ml-4">
