+++ vite.config.ts
@@ -1,8 +1,6 @@
+import { defineConfig } from 'vite';
 import { sveltekit } from '@sveltejs/kit/vite';
-import type { UserConfig } from 'vite';
-
-const config: UserConfig = {
-	plugins: [sveltekit()]
-};
-
-export default config;
+import tailwindcss from '@tailwindcss/vite';
+export default defineConfig({
+	plugins: [sveltekit(), tailwindcss()]
+});
